refactor: Move blocking API blocks to blocking_api.rs
